Move header methods in trait
